+++ UnityPackage/Runtime/Extensions/UnityEngine/Binding/UnityEngineBinding.cs
@@ -42,7 +42,6 @@ namespace EasyJection.Extensions
         public static Type TYPE_COMPONENT;
 
         private static IFactory GameObjectFactory;
-        private static IFactory ComponentFactory;
 
         static UnityEngineBinding()
         {
@@ -50,140 +49,28 @@ namespace EasyJection.Extensions
             TYPE_COMPONENT = typeof(Component);
             TYPE_GAMEOBJECT = typeof(GameObject);
             GameObjectFactory = new UnityEngineGameObjectInstantiateFactory();
-            ComponentFactory = new UnityEngineComponentInstantiateFactory();
         }
 
-        #region Comment
-        /// <summary>
-        /// Creates a binding of the key type to the <paramref name="type"/> on a GameObject of a given <paramref name="objectName"/> as a singleton.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// Recommend binding only to <see cref="UnityEngine.GameObject"/> that will not be destroyed in the scene to prevent references to destroyed objects.
-        /// </para>
-        /// <para>
-        /// 1️⃣ If the GameObject with the specified  <paramref name="objectName"/> is not found on the game scene, it will be added.
-        /// </para>
-        /// <para>
-        /// 2️⃣ If the GameObject with the specified  <paramref name="objectName"/> is not found on the GameObject, it will be added.
-        /// </para>
-        /// <para>
-        /// 3️⃣ If the <paramref name="type"/> is <see cref="UnityEngine.GameObject"/> binds the key to the GameObject itself.
-        /// </para>
-        /// <para>
-        /// 4️⃣ If the <paramref name="type"/> is <see cref="UnityEngine.Component"/> binds the key to an instance.
-        /// </para>
-        /// </remarks>
-        /// <param name="bindingFactory">The original binding factory.</param>
-        /// <param name="type">The component type.</param>
-        /// <param name="objectName">The GameObject name.</param>
-        /// <returns>The binding condition object related to this binding.</returns>
-        #endregion
-        public static IBindingInjection ToGameObject(this IBindingFactory bindingFactory, Type type)
+        public static IBindingInjection ToGameObject<T>(this IBindingFactory bindingFactory)
+        {
+            var type = typeof(T);
+            return bindingFactory.ToGameObject(type, type.Name);
+        }
+
+        public static IBindingInjection ToGameObject<T>(this IBindingFactory bindingFactory, string prefabKeyName)
+        {
+            return bindingFactory.ToGameObject(typeof(T), prefabKeyName);
+        }
+
+        public static IBindingInjection ToGameObject(this IBindingFactory bindingFactory, Type type, string gameobjectKey)
         {
             if (!Helper.IsAssignable(bindingFactory.BindingType, type))
                 throw new Exception(Causes.TYPE_NOT_ASSIGNABLE);
             if (type.IsInterface || type.IsAbstract)
                 throw new Exception(Causes.TYPE_NOT_IMPLEMENTED);
-            if (!Helper.IsAssignable(TYPE_GAMEOBJECT, type))
-                throw new Exception(string.Format("The \"{0}\" type must be derived from UnityEngine.GameObject.", type.Name));
 
-            return bindingFactory.AddFactoryInstance(type, GameObjectFactory);
+            return bindingFactory.AddFactoryInstance(type, GameObjectFactory, new PrefabBinding(gameobjectKey, type));
         }
-
-        public static IBindingInjection ToGameObject(this IBindingFactory bindingFactory) 
-        {
-            return bindingFactory.ToGameObject(bindingFactory.BindingType);
-        }
-
-        public static IBindingInjection ToComponent(this IBindingFactory bindingFactory, Type type)
-        {
-            if (!Helper.IsAssignable(bindingFactory.BindingType, type))
-                throw new Exception(Causes.TYPE_NOT_ASSIGNABLE);
-            if (type.IsInterface || type.IsAbstract)
-                throw new Exception(Causes.TYPE_NOT_IMPLEMENTED);
-            if (!Helper.IsAssignable(TYPE_COMPONENT, type))
-                throw new Exception(string.Format("The \"{0}\" type must be derived from UnityEngine.Component.", type.Name));
-
-            return bindingFactory.AddFactoryInstance(type, ComponentFactory);
-        }
-
-        //public static IBindingInjection ToGameObject(this IBindingFactory bindingFactory, Type type, string objectName = null)
-        //{
-
-        //    GameObject instance = null;
-
-        //    if (string.IsNullOrEmpty(objectName))
-        //        instance = new GameObject(type.Name);
-        //    else
-        //        instance = GameObject.Find(objectName) ?? new GameObject(type.Name);
-
-        //    return CreateSingletonProvider(bindingFactory, instance, type, objectType);
-        //}
-        /*
-        #region Comment
-        /// <summary>
-        /// Creates a binding of the key type to the <typeparamref name="T"/> type on a GameObject of a given <paramref name="objectName"/> as a singleton.
-        /// </summary>
-        /// <remarks>
-        /// <para>
-        /// If the GameObject with the specified  <paramref name="objectName"/> is not found on the GameObject, it will be added.
-        /// </para>
-        /// </remarks>
-        /// <typeparam name="T">The component type.</typeparam>
-        /// <param name="bindingFactory">The original binding factory.</param>
-        /// <param name="objectName">The GameObject name.</param>
-        /// <returns>The binding condition object related to this binding.</returns>
-        #endregion
-        public static IBindingInjection ToGameObject<T>(this IBindingFactory bindingFactory, string objectName) where T : Component
-        {
-            return bindingFactory.ToGameObject(typeof(T), objectName);
-        }
-
-        #region Comment
-        /// <summary>
-        /// Creates a binding of the key type to the <typeparamref name="T"/> type on a new GameObject as a singleton.
-        /// </summary>
-        /// <remarks>
-        /// </remarks>
-        /// <typeparam name="T">The component type.</typeparam>
-        /// <param name="bindingFactory">The original binding factory.</param>
-        /// <returns>The binding condition object related to this binding.</returns>
-        #endregion
-        public static IBindingInjection ToGameObject<T>(this IBindingFactory bindingFactory) where T : Component
-        {
-            return bindingFactory.ToGameObject(typeof(T));
-        }
-
-        #region Comment
-        /// <summary>
-        /// Creates a binding of the key type to itself on a new GameObject as a singleton.
-        /// ⚠ The key type must be derived either from <see cref="UnityEngine.GameObject"/> or <see cref="UnityEngine.Component"/>.
-        /// </summary>
-        /// <remarks>
-        /// Recommend binding only to <see cref="UnityEngine.GameObject"/> that will not be destroyed in the scene to prevent references to destroyed objects.
-        /// </remarks>
-        /// <param name="bindingFactory">The original binding factory.</param>
-        /// <returns>The binding condition object related to this binding.</returns>
-        #endregion
-        public static IBindingInjection ToGameObject(this IBindingFactory bindingFactory)
-        {
-            return bindingFactory.ToGameObject(bindingFactory.BindingType);
-        }
-
-        private static UnityEngineBindingCondition CreateSingletonProvider(IBindingFactory bindingFactory, GameObject gameObject, Type type, UnityEngineObjectType objectType)
-        {
-            if (gameObject == null)
-                throw new ArgumentNullException("The argument named 'gameObject' is null");
-
-            if (objectType == UnityEngineObjectType.GameObject)
-                return new UnityEngineBindingCondition(bindingFactory.AddBinding(gameObject, BindingInstanceType.Instance, false), gameObject.name);
-
-            var component = gameObject.GetComponent(type) ?? gameObject.AddComponent(type);
-
-            return new UnityEngineBindingCondition(bindingFactory.AddBinding(component, BindingInstanceType.Instance, false), gameObject.name);
-        }
-        */
     }
 #endif
 }