+++ UnityPackage/Samples/2_Binding Prefab/Prefabs/Scripts/IRotate.cs
@@ -1,0 +1,10 @@
+﻿// IRotate.cs
+using UnityEngine;
+
+namespace EasyJection.Samples.BindingPrefab
+{
+    public interface IRotate
+    {
+        void DoRotate(float x, float y, float z);
+    }
+}