+++ Framework/CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2023-02-12
+- Fixed bugs with dependency resolution for instances created by the factory.
+- Added support resolve for an array in field.
+
 ## [1.0.1] - 2023-02-11
 - Fixed a bug with injection to a default constructor.
 - Added the ability to bind a factory class without an inherited 'EasyJection.Types.IFactory' and specify the method for creating instances by method name.